Remove "techpreview" filter on verify job
`oc` is planning on introducing feature gated tests that would run
/against/ a normal cluster (i.e. a CI job that doesn't have
featureset=techpreview).  In order to collect these results, we need to
remove the filter on featureset=techpreview.

For other FG, the tests only appear in jobs that are techpreview, so I
believe removing this filter from Sippy's queries is a net neutral
change.