API changes to set IPFamily within PlatformStatus of Infra CR
This config is being added to be explicit about the IP address family
for the Cluster so that we don't have to infer that from other configs
that may not always be present.

Updating PlatformStatus for AWS, Azure and GCP to enable dual stack
installs on these platforms (initially in tech preview).