NO-JIRA: New rules about CO's Degraded and Available conditions
The essence of the new rules is that operators MUST not go Available=False
or Degraded=True in an HA cluster during an uneventful CI upgrade.

Those rules have applied in CI for a while [1, 2] and OCPBugs have been
filed in this area. In order to avoid CI failing, many exceptions have been
added in the tests [3, 4] as many of those bugs are still open. It is expected
to invest effort to deliver the fixes of those bugs.

[1]. https://issues.redhat.com/browse/OTA-700

[2]. https://issues.redhat.com/browse/TRT-1578

[3]. https://github.com/openshift/origin/blob/2af38a7807699b3046a73f931884152a11271d21/pkg/monitortests/clusterversionoperator/legacycvomonitortests/operators.go#L102

[4]. https://github.com/openshift/origin/pull/27231