Add more details about the Progressing condition
"version change" is counted as config changes and thus cluster
operators must report Progressing for it.

On the contrary, cluster operators should not report Progress ONLY when
it owns a DaemonSet that acts on a new node or a node rebooting.
It is because it may happen too often in a heaviy-scaled cluster and make
Progressing less useful for the clients that hope to use it for Cluster
version updating.