CNTRLPLANE-1540: Add support for goaway-chance in the Kube API Server Operator
This adds a new quantity (in lieu for a float64 setting in KAS) that
will control the HTTP2 goaway chance on the kube-apiserver.

This feature already exists with a default value of 0.001, backported with supportex to 4.18.
We take the exact same default value, just making it a first class
citizen on the apiserver cluster CRD.

Signed-off-by: Thomas Jungblut <tjungblu@redhat.com>