OCPEDGE-2084: Add PacemakerStatus CRD for Two-Node OpenShift with Fencing (TNF)
Introduces etcd.openshift.io/v1alpha1 API group with a PacemakerCluster
custom resource. This provides visibility into pacemaker cluster health for
Two Node OpenShift with Fencing (TNF) etcd deployments. This status-only resource
provides data into the health of the pacemaker nodes and the three major resources
that they manage - kubelet, etcd, and a redfish fencing agent.

The cluster-etcd-operator is responsible for interpreting this an ensuring that
it degrades appropriately when etcd is unhealthy or at risk of being unable to
recover automatically in the event of a quorum-loss event.